DonorSearcOCR/app/requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.3.1+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html
albucore==0.0.16
albumentations==1.4.11
Pillow==10.1.0
annotated-types==0.7.0
anyio==3.7.1
click==8.1.5
exceptiongroup==1.1.2
fastapi==0.110.0
h11==0.14.0
idna==3.4
pydantic==2.9.2
pydantic-core==2.23.4
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.36.3
timm==1.0.8
typing-extensions==4.12.2
uvicorn==0.22.0
numpy==1.26.4
